Set up asynchronous pingbacks for Discourse/Ore.
This avoids having the pingbacks inline, which means any issues
encountered won't block login nor will they fail pinging the other
service. In addition, retries can be done independently and out of band.

Note that the pingback arguments are deliberately set to just the name
of the service to ping as well as the user ID: this is because we don't
want to accidentally save a "bad" pingback that will overwrite a
later set of data. By storing just the information needed to fetch the
rest of the data, we avoid this issue since we'll always use the
freshest data as at request time.